jgz80: Use 16 bit addresses for IO port operations

+++ z80.h
@@ -23,8 +23,8 @@ typedef struct z80 z80;
 struct z80 {
   uint8_t (*read_byte)(void*, uint16_t);
   void (*write_byte)(void*, uint16_t, uint8_t);
-  uint8_t (*port_in)(z80*, uint8_t);
-  void (*port_out)(z80*, uint8_t, uint8_t);
+  uint8_t (*port_in)(z80*, uint16_t);
+  void (*port_out)(z80*, uint16_t, uint8_t);
   void* userdata;
 
   uint16_t pc, sp, ix, iy; // special purpose registers
